update feature file structure
